pyproject.toml
[tool.poetry]
name = "excourses-bot-tg-"
version = "0.1.0"
description = ""
authors = ["НаноботАртем2208 <astramoda30@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.7.0"
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
colorama = "^0.4.6"
pydantic-settings = "^2.3.2"
sqlalchemy = "^2.0.30"
redis = "^5.0.5"
fastapi-cache2 = "0.2.1"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
aiosqlite = "^0.20.0"
aiogram-inline-paginations = "^0.1.12"
opencv-python = "^4.10.0.84"
gunicorn = "^23.0.0"
docker = "^7.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
